@@ -10,27 +10,40 @@ function Carousel(props) {
   const card = props.cardData[cardIdx];
   const total = props.cardData.length;
   const goForward = () => setCardIdx(cardIdx + 1);
+  const goBackward = () => setCardIdx(cardIdx - 1);
+
+  const leftArrow = (
+    <i
+      className="fas fa-chevron-circle-left fa-2x"
+      onClick={goBackward}
+      data-testid="left-arrow"
+    />
+  )
+
+  const rightArrow = (
+    <i
+      className="fas fa-chevron-circle-right fa-2x"
+      onClick={goForward}
+      data-testid="right-arrow"
+    />
+  )
 
   return (
     <div className="Carousel">
       <h1>{props.title}</h1>
       <div className="Carousel-main">
-        <i
-          className="fas fa-chevron-circle-left fa-2x"
-          onClick={goForward}
-          data-testid="left-arrow"
-        />
+
+        {cardIdx === 0 ? null : leftArrow}
+        
         <Card
           caption={card.caption}
           src={card.src}
           currNum={cardIdx + 1}
           totalNum={total}
         />
-        <i
-          className="fas fa-chevron-circle-right fa-2x"
-          onClick={goForward}
-          data-testid="right-arrow"
-        />
+        
+        {cardIdx === total - 1 ? null : rightArrow}
+
       </div>
     </div>
   );
